backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
numpy==1.24.3
pandas==1.5.3
scikit-learn==1.0.2
sentence-transformers==2.2.0
PyPDF2==2.0.0
python-docx==0.8.11
requests==2.26.0
python-dotenv==0.19.0
joblib==1.1.0
scipy==1.7.3
matplotlib==3.5.3
seaborn==0.12.2
torch==1.9.0
transformers==4.11.0
tqdm==4.66.0
huggingface-hub==0.19.4
tokenizers==0.14.1
protobuf==4.25.1
regex==2023.10.3
safetensors==0.4.0
aiohttp==3.9.1
PyMuPDF==1.23.0
pdfplumber==0.10.3
nltk==3.6.3
spacy==3.1.0
python-magic==0.4.24
python-magic-bin==0.4.14
docx2txt==0.8
beautifulsoup4==4.9.3
lxml==4.9.0
pdfminer.six==20211012
gunicorn==20.1.0
werkzeug==2.0.1
pytest==6.2.5
pytest-cov==2.12.1
black==21.9b0
flake8==3.9.2
isort==5.9.3
mypy==0.910
urllib3==1.26.7
certifi==2023.7.22
cryptography==3.4.7
pyOpenSSL==20.0.1
google-generativeai==0.3.2 